Update README.md
This project is a good opportunity to work with input and output in Java. We will deal with arrays and functions as well as loops and conditional statements.

Cinema has been a part of the entertainment industry for a long time: a good movie is a way to escape reality and live through a variety of emotions. The best cinema experience you can get is probably in a cinema theatre. In this project, we will create an application that helps manage a cinema theatre: sell tickets, check available seats, see sales statistics, and more.